Simplify the HPX threading implementation.
